requirements.txt
transformers
datasets
sentence-transformers
chromadb
langchain
openai
ragas
streamlit
python-dotenv
google-generativeai 
langchain-google-genai
langchain_huggingface